+++ cli/tests/test_completion.rs
@@ -1559,6 +1559,7 @@ fn test_merge_tools() {
     let output = test_env.run_jj_in(dir, ["--", "jj", "resolve", "--tool", ""]);
     insta::assert_snapshot!(output, @"
     :builtin
+    :select
     :ours
     :theirs
     kdiff3